refactor(server): move connector authorization policy into the connectors package
GrantTypeAllowed, the connector-allowed check, and the ConnectorGrantTypes set
are pure connector-authorization policy shared by the token grants and the
browser auth flow. Move them into the connectors package (as GrantTypeAllowed,
ConnectorAllowed and ConnectorGrantTypes) so the grants package can enforce the
same policy directly, without reaching back into the server. No behaviour change.
