+++ cmd/dex/serve.go
@@ -38,6 +38,7 @@ import (
 	"github.com/dexidp/dex/api/v2"
 	"github.com/dexidp/dex/pkg/featureflags"
 	"github.com/dexidp/dex/server"
+	"github.com/dexidp/dex/server/connectors"
 	"github.com/dexidp/dex/server/signer"
 	"github.com/dexidp/dex/storage"
 )
@@ -257,7 +258,7 @@ func runServe(options serveOptions) error {
 			return fmt.Errorf("invalid config: no config field for connector %q", c.ID)
 		}
 		for _, gt := range c.GrantTypes {
-			if !server.ConnectorGrantTypes[gt] {
+			if !connectors.ConnectorGrantTypes[gt] {
 				return fmt.Errorf("invalid config: unknown grant type %q for connector %q", gt, c.ID)
 			}
 		}
